2020-1/pyopengl_test/requirements.txt
autopep8 == 1.5.3
glfw == 1.11.2
multipledispatch == 0.6.0
numpy == 1.18.5
Pillow == 7.1.2
pip == 20.1.1
pycodestyle == 2.6.0
pygame == 1.9.6
PyOpenGL == 3.1.5
pyrr == 0.10.3
setuptools ==41.2.0
six == 1.15.0
toml == 0.10.1
